map Map stuff to its own component
