+++ src/GameBoard.css
@@ -1,11 +1,3 @@
-#map {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
 #label {
   display: none;
   position: absolute;
